don't write obal extra vars if there are none

+++ puppet/modules/jenkins_job_builder/files/theforeman.org/pipelines/lib/obal.groovy
@@ -1,17 +1,28 @@
 def obal(args) {
-    def extra_vars_file = writeExtraVars(extraVars: args.extraVars)
+    def extra_vars = args.extraVars ?: [:]
+    def extra_vars_file
+
     def packages = args.packages
     if (packages instanceof String[]) {
         packages = packages.join(' ')
     }
 
+    def cmd = "python -m obal ${args.action} ${packages}"
+
+    if (extra_vars) {
+        extra_vars_file = writeExtraVars(extraVars: extra_vars)
+        cmd = "${cmd} -e @${extra_vars_file}"
+    }
+
     wrap([$class: 'AnsiColorBuildWrapper', colorMapName: "xterm"]) {
         withEnv(['ANSIBLE_FORCE_COLOR=true', "PYTHONPATH=${env.WORKSPACE}/obal"]) {
-            sh "python -m obal ${args.action} ${packages} -e @${extra_vars_file}"
+            sh "${cmd}"
         }
     }
 
-    sh "rm ${extra_vars_file}"
+    if (extra_vars) {
+        sh "rm ${extra_vars_file}"
+    }
 }
 
 def setup_obal() {
